+++ src/include/digit.h
@@ -4,21 +4,21 @@
 #include <Arduino.h>
 
 struct digitPinout {
-  int a;
-  int b;
-  int c;
-  int d;
-  int e;
-  int f;
-  int g;
-  int DP;
-  int BR;
+  uint8_t a;
+  uint8_t b;
+  uint8_t c;
+  uint8_t d;
+  uint8_t e;
+  uint8_t f;
+  uint8_t g;
+  uint8_t DP;
+  uint8_t BR;
 };
 
 extern struct digitPinout digitPin;
-extern int8_t digitEnc[];
+extern uint8_t digitEnc[];
 
 // Helpers
-void digShow(int digit, struct digitPinout* digp);
+void digShow(uint8_t digit, struct digitPinout* digp);
 
 #endif
